Moved evidently integration to correct folder, examples inside evidently folder
