--- src/coalescenceml/monitors/BaseProfileStep.py
@@ -1,3 +1,0 @@
-class BaseProfileStep():
-  def __init__(self) -> None:
-      pass
